+++ database/migrations/2017_06_14_200801_create_mimics_table.php
@@ -15,13 +15,13 @@ class CreateMimicsTable extends Migration
     {
         Schema::create('mimics', function (Blueprint $table) {
             $table->engine = 'InnoDB';
-            $table->integer('id', true);
+            $table->bigInteger('id', true);
             $table->string('media', 255);
             $table->tinyInteger('mimic_type'); //video/picture
             $table->tinyInteger('is_response'); //0/1 - is this mimic response or not
             $table->tinyInteger('is_private')->default(0); //0/1 - is this mimic private or not
-            $table->integer('upvote')->default(0);
-            $table->integer('user_id');
+            $table->bigInteger('upvote')->default(0);
+            $table->bigInteger('user_id');
             $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('restrict');
         });
