+++ database/migrations/2017_06_14_201138_create_hashtags_table.php
@@ -15,7 +15,7 @@ class CreateHashtagsTable extends Migration
     {
         Schema::create('hashtags', function (Blueprint $table) {
             $table->engine = 'InnoDB';
-            $table->integer('id', true);
+            $table->bigInteger('id', true);
             $table->string('name', 255);
             $table->integer('popularity');
             $table->timestamps();
