+++ database/migrations/2017_06_14_201151_create_mimic_hashtag_table.php
@@ -15,9 +15,9 @@ class CreateMimicHashtagTable extends Migration
     {
         Schema::create('mimic_hashtag', function (Blueprint $table) {
             $table->engine = 'InnoDB';
-            $table->integer('id', true);
-            $table->integer('mimic_id');
-            $table->integer('hashtag_id');
+            $table->bigInteger('id', true);
+            $table->bigInteger('mimic_id');
+            $table->bigInteger('hashtag_id');
             $table->timestamps();
             $table->foreign('mimic_id')->references('id')->on('mimics')->onUpdate('cascade')->onDelete('cascade');
             $table->foreign('hashtag_id')->references('id')->on('hashtags')->onUpdate('cascade')->onDelete('cascade');
